event_planner/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
celery==5.4.0
redis==5.2.0
eventlet==0.38.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
Pillow==11.0.0
openpyxl==3.1.5
pytest==8.3.5
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.3